debug: found that createFracBox() and potential other vsetMakers had buggy 'pfverts'; now this is recreated and problem fixed static analyser: fixed a number of issues identified by strict analysis such unsigned's that should have been signed integers etc chore: tidied up constructors for Face chore: removed some testing code from BoundaryInterface

+++ source/includes/interfaces/Eclipse/CellGenerator.h
@@ -270,7 +270,7 @@ public:
 
 	bool ConstructHexahedron(ColumnCell& cell) {
 		//std::cerr << "hexa: \n";				
-		for ( size_t i = 0; i < 8; ++i) {
+		for ( uint32_t i = 0; i < 8; ++i) {
 			vertexIDs[i] = i;
 			auto p = getNodeCoord(cell, vertexIDs[i]);
 			grid.ConvertFromReservoirToCSMPcoordinateSystem(p);
