+++ source/modules/interfaces/GeoModeller/IntrepidInterface.cpp
@@ -369,7 +369,7 @@ void IntrepidInterface::RepairElementOrientations( VSet<3U>& vset ) const
             if ( tet.Volume() < 0. ) {
                 tet.CounterClockwiseNodes( ids );
                 // reassigning the nodes to the plist in opposite order (first getting the global node numbers
-                for ( auto it=ids.begin(); it!=ids.end(); ++it ) (*it) = vset.Plist( elmt, (*it) );
+                for ( auto it=ids.begin(); it!=ids.end(); ++it ) (*it) = static_cast<uint32_t>( vset.Plist( elmt, (*it) ));
                 size_t counter(0U);
                 for ( auto nit=vset.PlistBegin(elmt); nit!=vset.PlistEnd(elmt); ++nit )
                   (*nit) = ids[counter++];
